Circle fixes: messages async initial state
