+++ src/ui/setup/index.ts
@@ -14,13 +14,19 @@ import { getUserSettings } from "ui/hooks/settings";
 import { initLaunchDarkly } from "ui/utils/launchdarkly";
 import { maybeSetMixpanelContext } from "ui/utils/mixpanel";
 import { syncInitialLayoutState } from "ui/reducers/layout";
-import { initialMessageState } from "devtools/client/webconsole/reducers/messages";
+import {
+  FiltersState,
+  MessageState,
+  syncInitialMessageState,
+  defaultFiltersState,
+} from "devtools/client/webconsole/reducers/messages";
 import { trackEvent } from "ui/utils/telemetry";
 import { Recording } from "ui/types";
 import { getRecording } from "ui/hooks/recordings";
 import { getRecordingId } from "ui/utils/recording";
 import { getReplaySession } from "ui/setup/prefs";
 import type { LayoutState } from "ui/state/layout";
+import { getLocalReplaySessionPrefs } from "ui/setup/prefs";
 import type { TabsState } from "devtools/client/debugger/src/reducers/tabs";
 import { EMPTY_TABS } from "devtools/client/debugger/src/reducers/tabs";
 import { CommentsState } from "ui/state/comments";
@@ -114,6 +120,25 @@ export async function getInitialCommentsState(): Promise<CommentsState> {
     pendingComment: session?.pendingComment || null,
   };
 }
+
+const getInitialFiltersState = async () => {
+  const session = await getLocalReplaySessionPrefs();
+
+  return session ? { ...defaultFiltersState, ...session.consoleFilters } : defaultFiltersState;
+};
+
+export const initialMessageState = async (
+  overrides: Partial<MessageState> = {}
+): Promise<MessageState> => {
+  // Realistically, we only expect filters and commandHistory
+  // See ui/setup/dynamic/devtools.ts
+  const { filters = {}, ...otherOverrides } = overrides;
+
+  return syncInitialMessageState({
+    ...overrides,
+    filters: { ...(await getInitialFiltersState()), ...filters },
+  });
+};
 
 export async function bootstrapApp() {
   const initialState = {
